+++ internal/sandbox/cell_timeout_test.go
@@ -23,6 +23,9 @@ func (r *recordingImpl) runPython(_ context.Context, req PythonRequest) (PythonR
 func (r *recordingImpl) resourceUsage() (ResourceUsageSummary, bool) {
 	return ResourceUsageSummary{}, false
 }
+func (r *recordingImpl) runFileOp(context.Context, FileOpRequest) (FileOpResult, error) {
+	return FileOpResult{}, nil
+}
 func (r *recordingImpl) poisoned() bool { return false }
 func (r *recordingImpl) close()         {}